+++ src/pvr_client_base.cpp
@@ -794,7 +794,7 @@ int PVRClientBase::ReadLiveStream(unsigned char* pBuffer, unsigned int iBufferSi
     return bytesRead;
 }
 
-long long PVRClientBase::SeekLiveStream(long long iPosition, int iWhence)
+int64_t PVRClientBase::SeekLiveStream(int64_t iPosition, int iWhence)
 {
     CLockObject lock(m_mutex);
     if(nullptr == m_inputBuffer){
@@ -803,7 +803,7 @@ long long PVRClientBase::SeekLiveStream(long long iPosition, int iWhence)
     return m_inputBuffer->Seek(iPosition, iWhence);
 }
 
-long long PVRClientBase::PositionLiveStream()
+int64_t PVRClientBase::PositionLiveStream()
 {
     CLockObject lock(m_mutex);
     if(nullptr == m_inputBuffer){
@@ -812,7 +812,7 @@ long long PVRClientBase::PositionLiveStream()
     return m_inputBuffer->GetPosition();
 }
 
-long long PVRClientBase::LengthLiveStream()
+int64_t PVRClientBase::LengthLiveStream()
 {
     CLockObject lock(m_mutex);
     if(nullptr == m_inputBuffer){
@@ -1310,16 +1310,16 @@ int PVRClientBase::ReadRecordedStream(unsigned char *pBuffer, unsigned int iBuff
     return (m_recordBuffer.buffer == NULL) ? -1 : m_recordBuffer.buffer->Read(pBuffer, iBufferSize, (m_recordBuffer.isLocal)? 0 : ChannelReloadTimeout() * 1000);
 }
 
-long long PVRClientBase::SeekRecordedStream(long long iPosition, int iWhence)
+int64_t PVRClientBase::SeekRecordedStream(int64_t iPosition, int iWhence)
 {
     return (m_recordBuffer.buffer == NULL) ? -1 : m_recordBuffer.buffer->Seek(iPosition, iWhence);
 }
 
-long long PVRClientBase::PositionRecordedStream(void)
+int64_t PVRClientBase::PositionRecordedStream(void)
 {
     return (m_recordBuffer.buffer == NULL) ? -1 : m_recordBuffer.buffer->GetPosition();
 }
-long long PVRClientBase::LengthRecordedStream(void)
+int64_t PVRClientBase::LengthRecordedStream(void)
 {
     return (m_recordBuffer.buffer == NULL) ? -1 : m_recordBuffer.buffer->GetLength();
 }
